+++ CHANGES.md
@@ -2,6 +2,13 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/).
 
+## 9.7.0
+
+- Added optional entry character to the **Parse IDS** feature of the **Unihan Components** utility.
+- Updated `IDS.TXT` data file.
+- Updated `BabelStoneHanPUA.woff2` font file.
+- Updated `Electron` to version `12.0.4`.
+
 ## 9.6.0
 
 - Added three Unihan sample scripts: `List IDS Indivisible Characters`, `List IDS Unrepresentable Characters`, and `Detect Radical-Strokes Discrepancies` to the **JavaScript Runner** utility.